Move custom body generation into generators
