+++ kybra/compiler/kybra_generate/src/py_ast/mod.rs
@@ -1,21 +1,18 @@
-use quote::quote;
-use std::collections::{HashMap, HashSet};
-
 use cdk_framework::{
     nodes::{act_canister_method, data_type_nodes, ActExternalCanister},
     ActCanisterMethod, ActDataType, CanisterMethodType,
 };
+use std::collections::{HashMap, HashSet};
 
-use crate::generators::{
-    async_result_handler::generate_async_result_handler, kybra_serde::generate_kybra_serde,
-    rng_seed, stable_b_tree_map::generate_stable_b_tree_map,
+use crate::{
+    generators::body,
+    py_ast::{
+        kybra_types::{KybraStmt, StableBTreeMapNode},
+        traits::GetDependencies,
+    },
 };
 
 pub use self::{kybra_ast::KybraAst, kybra_types::KybraProgram};
-use self::{
-    kybra_types::{KybraStmt, StableBTreeMapNode},
-    traits::GetDependencies,
-};
 
 mod kybra_ast;
 mod system_methods;
@@ -101,37 +98,7 @@ impl PyAst<'_> {
 
         let stable_b_tree_map_nodes = self.build_stable_b_tree_map_nodes();
 
-        let async_result_handler = generate_async_result_handler(&external_canisters);
-
-        let kybra_serde = generate_kybra_serde();
-
-        let stable_b_tree_map = generate_stable_b_tree_map(&stable_b_tree_map_nodes);
-
-        let rng_seed = rng_seed::generate();
-
-        let rust_code = quote! {
-            pub fn _kybra_unwrap_rust_python_result<T>(
-                rust_python_result: Result<T, PyRef<PyBaseException>>,
-                vm: &rustpython::vm::VirtualMachine
-            ) -> T {
-                match rust_python_result {
-                    Ok(ok) => ok,
-                    Err(err) => {
-                        let err_string: String = err.to_pyobject(vm).repr(vm).unwrap().to_string();
-
-                        panic!("{}", err_string);
-                    },
-                }
-            }
-
-            #async_result_handler
-
-            #kybra_serde
-
-            #stable_b_tree_map
-
-            #rng_seed
-        };
+        let rust_code = body::generate(&external_canisters, &stable_b_tree_map_nodes);
 
         KybraAst {
             init_method: self.build_init_method(
